+++ common/src/main/java/net/mehvahdjukaar/supplementaries/ClientProxy.java
@@ -1,5 +1,7 @@
 package net.mehvahdjukaar.supplementaries;
 
+import net.mehvahdjukaar.moonlight.api.platform.PlatHelper;
+import net.mehvahdjukaar.moonlight.api.platform.RegHelper;
 import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Player;