+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/misc/songs/SongsManager.java
@@ -19,6 +19,7 @@ import net.mehvahdjukaar.supplementaries.common.utils.MiscUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.sounds.SoundInstance;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.resources.ResourceManager;
@@ -53,12 +54,9 @@ public class SongsManager extends SimpleJsonResourceReloadListener {
 
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().disableHtmlEscaping().create();
 
-    public static final SongsManager INSTANCE = new SongsManager();
-
-    public SongsManager() {
+    public SongsManager(HolderLookup.Provider registryAccess) {
         super(GSON, "flute_songs");
     }
-
 
     @Override
     protected void apply(Map<ResourceLocation, JsonElement> jsons, ResourceManager manager, ProfilerFiller profile) {