+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/integration/fabric/AppleSkinCompatImpl.java
@@ -1,5 +1,7 @@
 package net.mehvahdjukaar.supplementaries.integration.fabric;
 
+import net.mehvahdjukaar.moonlight.fabric.DataMapBridge;
+
 public class AppleSkinCompatImpl {
     public static void addTooltipImage() {
     }