+++ neoforge/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/neoforge/ModSlider.java
@@ -4,6 +4,7 @@ import net.mehvahdjukaar.supplementaries.client.screens.widgets.ISlider;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.network.chat.Component;
 import net.neoforged.neoforge.client.gui.widget.ExtendedSlider;
+import net.neoforged.neoforge.common.Tags;
 
 public class ModSlider extends ExtendedSlider implements ISlider, GuiEventListener {
     public ModSlider(int x, int y, int width, int height, Component prefix, Component suffix, double minValue, double maxValue, double currentValue, double stepSize, int precision, boolean drawString) {