+++ changelog.md
@@ -1,3 +1,3 @@
 CHECK 3.0.0 PATCH NOTES!!
 
-- fixed a crash with flax
+- slidy block on honey